requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
datasets>=2.12.0
peft>=0.4.0
bitsandbytes>=0.41.0
accelerate>=0.20.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.25.0
plotly>=5.15.0
tqdm>=4.65.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
